refactor: introduce Frame type for type-safe frame numbering (#127)
* refactor: introduce Frame type for type-safe frame numbering

Define `type Frame uint` with Frame(0) as FrameForever (zero value =
unset/forever) and valid frames starting at 1. This replaces raw
uint/int fields across all core structs, parsers, storage backends,
and tests. The v1 JSON export converts FrameForever(0) back to -1
at the boundary for backwards compatibility.

* fix: convert 1-based frames to 0-based in v1 export boundary

The v1 JSON format is a stable contract with the JS frontend that uses
0-based frame numbering. Add frameToV1() helper that subtracts 1 from
all internal Frame values at the export boundary, preserving backwards
compatibility for existing recordings and new ones alike.

* fix: frameToV1 returns int so FrameForever(0) maps to -1

Changed frameToV1 return type from uint to int. This allows
Frame(0)/FrameForever to naturally map to -1 (the v1 sentinel
for "forever"), eliminating the separate FrameForever special
case for marker EndFrame. Updated all test assertions to match
the int return type in []any slices.

* chore: remove redundant int() cast around frameToV1

* refactor: use int for v1 export frame fields, remove uint() casts

The v1 Export struct fields (EndFrame, FrameNum, StartFrameNum) are now
int to match frameToV1's return type directly. This eliminates all
uint() wrapper casts and removes the risk of silent wraparound if
FrameForever ever reaches a struct field.

* fix: remove maxFrame guard so empty missions export EndFrame -1

Empty missions now get EndFrame: -1 (from frameToV1(FrameForever)),
distinguishing them from 1-frame missions (EndFrame: 0).

+++ internal/storage/postgres/postgres.go
@@ -271,7 +271,7 @@ func (b *Backend) DeleteMarker(dm *core.DeleteMarker) error {
 
 	deleteState := model.MarkerState{
 		MarkerID:     markerID,
-		CaptureFrame: dm.EndFrame,
+		CaptureFrame: uint(dm.EndFrame),
 		Time:         time.Now(),
 		Alpha:        0,
 	}
@@ -333,7 +333,7 @@ func (b *Backend) RecordRadioEvent(e *core.RadioEvent) error {
 func (b *Backend) RecordTelemetryEvent(e *core.TelemetryEvent) error {
 	b.queues.FpsEvents.Push(model.ServerFpsEvent{
 		Time:         e.Time,
-		CaptureFrame: e.CaptureFrame,
+		CaptureFrame: uint(e.CaptureFrame),
 		FpsAverage:   e.FpsAverage,
 		FpsMin:       e.FpsMin,
 	})
